+++ packages/openapi_base/lib/openapi_base.dart
@@ -1,3 +1,5 @@
+export 'dart:typed_data';
+
 export 'src/http_headers.dart';
 export 'src/openapi_base.dart';
 export 'src/openapi_client_base.dart';
@@ -6,4 +8,6 @@ export 'src/openapi_exception.dart';
 export 'src/server/openapi_server_base.dart';
 export 'src/server/openapi_shelf_noop_server.dart'
     if (dart.library.io) 'src/server/openapi_shelf_server.dart';
+export 'src/util/uint8list.dart';
 export 'src/util/uuid.dart';
+export 'src/server/openapi_shelf_server.dart' show ShelfRequest;