+++ packages/openapi_code_builder/example/test/build_verify_test.dart
@@ -1,0 +1,11 @@
+import 'package:build_verify/build_verify.dart';
+import 'package:test/test.dart';
+
+void main() {
+  test(
+    'ensure_build',
+    () {
+      expectBuildClean(packageRelativeDirectory: 'example');
+    },
+  );
+}