+++ packages/openapi_base/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.3.1+1
+
+* Updated package deps
+
 ## 1.3.1
 
 * bump UUID version constraint to support `4.x`
@@ -5,6 +9,10 @@
 ## 1.3.0
 
 * `ResponseMap` now has a return type.
+
+## 1.1.1
+
+* string binary support
 
 ## 1.1.0+3